refactor: expose group-state descriptor boundary
